Adds routes for classroom activations
- Added route for students to get a classroom's activation status
- Added route for teacher to activate a classroom
- Added route for teacher to deactivate a classroom
- refactored cors options into its own file
